[IMP] base, account: `refs` for multiple records retrieval
The `ref` method both in Environment and in ChartTemplate can now
accept multiple xml_ids, reducing the number of queries necessary.
The xml_ids must belong all to the same model or ValueError is raised.
`raise_if_not_found` argument is made a keyword-only argument so now
it's mandatory for the function calls must include the argument name.